Move options path and improve type inference
